+++ src/uet/oop/bomberman/Playground.java
@@ -4,18 +4,16 @@ import javafx.scene.Group;
 import javafx.scene.Scene;
 import javafx.scene.canvas.Canvas;
 import javafx.scene.canvas.GraphicsContext;
-import uet.oop.bomberman.Keyboard.Keyboard;
-import uet.oop.bomberman.entities.Bomber;
-import uet.oop.bomberman.entities.Character;
+import uet.oop.bomberman.keyboard.Keyboard;
+import uet.oop.bomberman.entities.character.Bomber;
+import uet.oop.bomberman.entities.character.Character;
 import uet.oop.bomberman.entities.Entity;
 import uet.oop.bomberman.entities.EntityGroup;
-import uet.oop.bomberman.entities.bomb.Bomb;
-import uet.oop.bomberman.entities.bomb.Flame;
-import uet.oop.bomberman.entities.tile.Tile;
+import uet.oop.bomberman.entities.character.bomb.Bomb;
+import uet.oop.bomberman.entities.character.bomb.Flame;
 import uet.oop.bomberman.graphics.Sprite;
 import uet.oop.bomberman.map.FileMapLoader;
 import uet.oop.bomberman.map.MapLoader;
-import uet.oop.bomberman.untility.Convert;
 import uet.oop.bomberman.untility.Point;
 
 import java.util.ArrayList;
