+++ src/uet/oop/bomberman/entities/Entity.java
@@ -2,14 +2,23 @@ package uet.oop.bomberman.entities;
 
 import javafx.scene.canvas.GraphicsContext;
 import javafx.scene.image.Image;
-import uet.oop.bomberman.entities.bomb.Flame;
 import uet.oop.bomberman.graphics.Sprite;
+import uet.oop.bomberman.untility.Convert;
+import uet.oop.bomberman.untility.Point;
 
 public abstract class Entity {
-
-    private int x;
-    private int y;
+    private int x; // in pixel
+    private int y; // in pixel
+    private Point coordinate; // in tile
     private Image spriteImg;
+
+    public Point getCoordinate() {
+        return coordinate;
+    }
+
+    public void setCoordinate(Point coordinate) {
+        this.coordinate = coordinate;
+    }
 
     public int getX() {
         return x;
@@ -17,6 +26,7 @@ public abstract class Entity {
 
     public void setX(int x) {
         this.x = x;
+        this.setCoordinate(Convert.pixelToTile(new Point(this.x, this.y)));
     }
 
     public int getY() {
@@ -25,6 +35,7 @@ public abstract class Entity {
 
     public void setY(int y) {
         this.y = y;
+        this.setCoordinate(Convert.pixelToTile(new Point(this.x, this.y)));
     }
 
     public Image getSpriteImg() {
@@ -39,13 +50,13 @@ public abstract class Entity {
      * Constructor.
      * Create new Entity with position and it's image
      *
-     * @param xUnit     position in predefined
-     * @param yUnit     position in predefined
+     * @param crdX      in tile
+     * @param crdY      in tile
      * @param spriteImg image
      */
-    public Entity(int xUnit, int yUnit, Image spriteImg) {
-        this.setX(xUnit * Sprite.SCALED_SIZE);
-        this.setY(yUnit * Sprite.SCALED_SIZE);
+    public Entity(int crdX, int crdY, Image spriteImg) {
+        this.setX(crdX * Sprite.SCALED_SIZE);
+        this.setY(crdY * Sprite.SCALED_SIZE);
 
         this.setSpriteImg(spriteImg);
     }
@@ -63,6 +74,4 @@ public abstract class Entity {
      * Update.
      */
     public abstract void update();
-
-    protected abstract boolean collide(Entity entity);
 }
