--- src/uet/oop/bomberman/entities/bomb/FramSegment.java
@@ -1,4 +1,0 @@
-package uet.oop.bomberman.entities.bomb;
-
-public class FramSegment {
-}
