+++ src/uet/oop/bomberman/entities/character/bomb/FramSegment.java
@@ -1,0 +1,4 @@
+package uet.oop.bomberman.entities.character.bomb;
+
+public class FramSegment {
+}
