+++ src/uet/oop/bomberman/graphics/Sprite.java
@@ -2,236 +2,42 @@ package uet.oop.bomberman.graphics;
 
 import javafx.scene.image.*;
 
-import java.util.Arrays;
-
 /**
- * Lưu trữ thông tin các pixel của 1 sprite (hình ảnh game)
+ * Manage a special sprite.
  */
-public final class Sprite {
-    // Constant class value
-    public static final int DEFAULT_SIZE = 16;
-    public static final int SCALED_SIZE = DEFAULT_SIZE * 2;
+public final class Sprite implements SpritePack {
     private static final int TRANSPARENT_COLOR = 0xffff00ff;
-
-    // Object value
     private final int size;
-    private int x, y;
+    private final int x, y;
     public int[] pixels;
-    protected int realWidth;
-    protected int realHeight;
-    private SpriteSheet sheet;
-
-    /*
-    |--------------------------------------------------------------------------
-    | Board sprites
-    |--------------------------------------------------------------------------
-     */
-    public static final Sprite grass = new Sprite(DEFAULT_SIZE, 6, 0, SpriteSheet.tiles, 16, 16);
-    public static final Sprite brick = new Sprite(DEFAULT_SIZE, 7, 0, SpriteSheet.tiles, 16, 16);
-    public static final Sprite wall = new Sprite(DEFAULT_SIZE, 5, 0, SpriteSheet.tiles, 16, 16);
-    public static final Sprite portal = new Sprite(DEFAULT_SIZE, 4, 0, SpriteSheet.tiles, 14, 14);
-
-    /*
-    |--------------------------------------------------------------------------
-    | Bomber Sprites
-    |--------------------------------------------------------------------------
-     */
-
-    public static final Sprite player_up = new Sprite(DEFAULT_SIZE, 0, 0, SpriteSheet.tiles, 12, 16);
-    public static final Sprite player_up_1 = new Sprite(DEFAULT_SIZE, 0, 1, SpriteSheet.tiles, 12, 16);
-    public static final Sprite player_up_2 = new Sprite(DEFAULT_SIZE, 0, 2, SpriteSheet.tiles, 12, 15);
-
-    public static final Sprite player_down = new Sprite(DEFAULT_SIZE, 2, 0, SpriteSheet.tiles, 12, 15);
-    public static final Sprite player_down_1 = new Sprite(DEFAULT_SIZE, 2, 1, SpriteSheet.tiles, 12, 15);
-    public static final Sprite player_down_2 = new Sprite(DEFAULT_SIZE, 2, 2, SpriteSheet.tiles, 12, 16);
-
-    public static final Sprite player_left = new Sprite(DEFAULT_SIZE, 3, 0, SpriteSheet.tiles, 10, 15);
-    public static final Sprite player_left_1 = new Sprite(DEFAULT_SIZE, 3, 1, SpriteSheet.tiles, 11, 16);
-    public static final Sprite player_left_2 = new Sprite(DEFAULT_SIZE, 3, 2, SpriteSheet.tiles, 12, 16);
-
-    public static final Sprite player_right = new Sprite(DEFAULT_SIZE, 1, 0, SpriteSheet.tiles, 10, 16);
-    public static final Sprite player_right_1 = new Sprite(DEFAULT_SIZE, 1, 1, SpriteSheet.tiles, 11, 16);
-    public static final Sprite player_right_2 = new Sprite(DEFAULT_SIZE, 1, 2, SpriteSheet.tiles, 12, 16);
-
-    public static final Sprite player_dead1 = new Sprite(DEFAULT_SIZE, 4, 2, SpriteSheet.tiles, 14, 16);
-    public static final Sprite player_dead2 = new Sprite(DEFAULT_SIZE, 5, 2, SpriteSheet.tiles, 13, 15);
-    public static final Sprite player_dead3 = new Sprite(DEFAULT_SIZE, 6, 2, SpriteSheet.tiles, 16, 16);
-
-    /*
-    |--------------------------------------------------------------------------
-    | Character
-    |--------------------------------------------------------------------------
-     */
-    //BALLOM
-    public static final Sprite balloom_left1 = new Sprite(DEFAULT_SIZE, 9, 0, SpriteSheet.tiles, 16, 16);
-    public static final Sprite balloom_left2 = new Sprite(DEFAULT_SIZE, 9, 1, SpriteSheet.tiles, 16, 16);
-    public static final Sprite balloom_left3 = new Sprite(DEFAULT_SIZE, 9, 2, SpriteSheet.tiles, 16, 16);
-
-    public static final Sprite balloom_right1 = new Sprite(DEFAULT_SIZE, 10, 0, SpriteSheet.tiles, 16, 16);
-    public static final Sprite balloom_right2 = new Sprite(DEFAULT_SIZE, 10, 1, SpriteSheet.tiles, 16, 16);
-    public static final Sprite balloom_right3 = new Sprite(DEFAULT_SIZE, 10, 2, SpriteSheet.tiles, 16, 16);
-
-    public static final Sprite balloom_dead = new Sprite(DEFAULT_SIZE, 9, 3, SpriteSheet.tiles, 16, 16);
-
-    //ONEAL
-    public static final Sprite oneal_left1 = new Sprite(DEFAULT_SIZE, 11, 0, SpriteSheet.tiles, 16, 16);
-    public static final Sprite oneal_left2 = new Sprite(DEFAULT_SIZE, 11, 1, SpriteSheet.tiles, 16, 16);
-    public static final Sprite oneal_left3 = new Sprite(DEFAULT_SIZE, 11, 2, SpriteSheet.tiles, 16, 16);
-
-    public static final Sprite oneal_right1 = new Sprite(DEFAULT_SIZE, 12, 0, SpriteSheet.tiles, 16, 16);
-    public static final Sprite oneal_right2 = new Sprite(DEFAULT_SIZE, 12, 1, SpriteSheet.tiles, 16, 16);
-    public static final Sprite oneal_right3 = new Sprite(DEFAULT_SIZE, 12, 2, SpriteSheet.tiles, 16, 16);
-
-    public static final Sprite oneal_dead = new Sprite(DEFAULT_SIZE, 11, 3, SpriteSheet.tiles, 16, 16);
-
-    //Doll
-    public static final Sprite doll_left1 = new Sprite(DEFAULT_SIZE, 13, 0, SpriteSheet.tiles, 16, 16);
-    public static final Sprite doll_left2 = new Sprite(DEFAULT_SIZE, 13, 1, SpriteSheet.tiles, 16, 16);
-    public static final Sprite doll_left3 = new Sprite(DEFAULT_SIZE, 13, 2, SpriteSheet.tiles, 16, 16);
-
-    public static final Sprite doll_right1 = new Sprite(DEFAULT_SIZE, 14, 0, SpriteSheet.tiles, 16, 16);
-    public static final Sprite doll_right2 = new Sprite(DEFAULT_SIZE, 14, 1, SpriteSheet.tiles, 16, 16);
-    public static final Sprite doll_right3 = new Sprite(DEFAULT_SIZE, 14, 2, SpriteSheet.tiles, 16, 16);
-
-    public static final Sprite doll_dead = new Sprite(DEFAULT_SIZE, 13, 3, SpriteSheet.tiles, 16, 16);
-
-    //Minvo
-    public static final Sprite minvo_left1 = new Sprite(DEFAULT_SIZE, 8, 5, SpriteSheet.tiles, 16, 16);
-    public static final Sprite minvo_left2 = new Sprite(DEFAULT_SIZE, 8, 6, SpriteSheet.tiles, 16, 16);
-    public static final Sprite minvo_left3 = new Sprite(DEFAULT_SIZE, 8, 7, SpriteSheet.tiles, 16, 16);
-
-    public static final Sprite minvo_right1 = new Sprite(DEFAULT_SIZE, 9, 5, SpriteSheet.tiles, 16, 16);
-    public static final Sprite minvo_right2 = new Sprite(DEFAULT_SIZE, 9, 6, SpriteSheet.tiles, 16, 16);
-    public static final Sprite minvo_right3 = new Sprite(DEFAULT_SIZE, 9, 7, SpriteSheet.tiles, 16, 16);
-
-    public static final Sprite minvo_dead = new Sprite(DEFAULT_SIZE, 8, 8, SpriteSheet.tiles, 16, 16);
-
-    //Kondoria
-    public static final Sprite kondoria_left1 = new Sprite(DEFAULT_SIZE, 10, 5, SpriteSheet.tiles, 16, 16);
-    public static final Sprite kondoria_left2 = new Sprite(DEFAULT_SIZE, 10, 6, SpriteSheet.tiles, 16, 16);
-    public static final Sprite kondoria_left3 = new Sprite(DEFAULT_SIZE, 10, 7, SpriteSheet.tiles, 16, 16);
-
-    public static final Sprite kondoria_right1 = new Sprite(DEFAULT_SIZE, 11, 5, SpriteSheet.tiles, 16, 16);
-    public static final Sprite kondoria_right2 = new Sprite(DEFAULT_SIZE, 11, 6, SpriteSheet.tiles, 16, 16);
-    public static final Sprite kondoria_right3 = new Sprite(DEFAULT_SIZE, 11, 7, SpriteSheet.tiles, 16, 16);
-
-    public static final Sprite kondoria_dead = new Sprite(DEFAULT_SIZE, 10, 8, SpriteSheet.tiles, 16, 16);
-
-    //ALL
-    public static final Sprite mob_dead1 = new Sprite(DEFAULT_SIZE, 15, 0, SpriteSheet.tiles, 16, 16);
-    public static final Sprite mob_dead2 = new Sprite(DEFAULT_SIZE, 15, 1, SpriteSheet.tiles, 16, 16);
-    public static final Sprite mob_dead3 = new Sprite(DEFAULT_SIZE, 15, 2, SpriteSheet.tiles, 16, 16);
-
-    /*
-    |--------------------------------------------------------------------------
-    | Bomb Sprites
-    |--------------------------------------------------------------------------
-     */
-    public static final Sprite bomb = new Sprite(DEFAULT_SIZE, 0, 3, SpriteSheet.tiles, 15, 15);
-    public static final Sprite bomb_1 = new Sprite(DEFAULT_SIZE, 1, 3, SpriteSheet.tiles, 13, 15);
-    public static final Sprite bomb_2 = new Sprite(DEFAULT_SIZE, 2, 3, SpriteSheet.tiles, 12, 14);
-
-    /*
-    |--------------------------------------------------------------------------
-    | FlameSegment Sprites
-    |--------------------------------------------------------------------------
-     */
-    public static final Sprite bomb_exploded = new Sprite(DEFAULT_SIZE, 0, 4, SpriteSheet.tiles, 16, 16);
-    public static final Sprite bomb_exploded1 = new Sprite(DEFAULT_SIZE, 0, 5, SpriteSheet.tiles, 16, 16);
-    public static final Sprite bomb_exploded2 = new Sprite(DEFAULT_SIZE, 0, 6, SpriteSheet.tiles, 16, 16);
-
-    public static final Sprite explosion_vertical = new Sprite(DEFAULT_SIZE, 1, 5, SpriteSheet.tiles, 16, 16);
-    public static final Sprite explosion_vertical1 = new Sprite(DEFAULT_SIZE, 2, 5, SpriteSheet.tiles, 16, 16);
-    public static final Sprite explosion_vertical2 = new Sprite(DEFAULT_SIZE, 3, 5, SpriteSheet.tiles, 16, 16);
-
-    public static final Sprite explosion_horizontal = new Sprite(DEFAULT_SIZE, 1, 7, SpriteSheet.tiles, 16, 16);
-    public static final Sprite explosion_horizontal1 = new Sprite(DEFAULT_SIZE, 1, 8, SpriteSheet.tiles, 16, 16);
-    public static final Sprite explosion_horizontal2 = new Sprite(DEFAULT_SIZE, 1, 9, SpriteSheet.tiles, 16, 16);
-
-    public static final Sprite explosion_horizontal_left_last = new Sprite(DEFAULT_SIZE, 0, 7, SpriteSheet.tiles, 16, 16);
-    public static final Sprite explosion_horizontal_left_last1 = new Sprite(DEFAULT_SIZE, 0, 8, SpriteSheet.tiles, 16, 16);
-    public static final Sprite explosion_horizontal_left_last2 = new Sprite(DEFAULT_SIZE, 0, 9, SpriteSheet.tiles, 16, 16);
-
-    public static final Sprite explosion_horizontal_right_last = new Sprite(DEFAULT_SIZE, 2, 7, SpriteSheet.tiles, 16, 16);
-    public static final Sprite explosion_horizontal_right_last1 = new Sprite(DEFAULT_SIZE, 2, 8, SpriteSheet.tiles, 16, 16);
-    public static final Sprite explosion_horizontal_right_last2 = new Sprite(DEFAULT_SIZE, 2, 9, SpriteSheet.tiles, 16, 16);
-
-    public static final Sprite explosion_vertical_top_last = new Sprite(DEFAULT_SIZE, 1, 4, SpriteSheet.tiles, 16, 16);
-    public static final Sprite explosion_vertical_top_last1 = new Sprite(DEFAULT_SIZE, 2, 4, SpriteSheet.tiles, 16, 16);
-    public static final Sprite explosion_vertical_top_last2 = new Sprite(DEFAULT_SIZE, 3, 4, SpriteSheet.tiles, 16, 16);
-
-    public static final Sprite explosion_vertical_down_last = new Sprite(DEFAULT_SIZE, 1, 6, SpriteSheet.tiles, 16, 16);
-    public static final Sprite explosion_vertical_down_last1 = new Sprite(DEFAULT_SIZE, 2, 6, SpriteSheet.tiles, 16, 16);
-    public static final Sprite explosion_vertical_down_last2 = new Sprite(DEFAULT_SIZE, 3, 6, SpriteSheet.tiles, 16, 16);
-
-    /*
-    |--------------------------------------------------------------------------
-    | Brick FlameSegment
-    |--------------------------------------------------------------------------
-     */
-    public static final Sprite brick_exploded = new Sprite(DEFAULT_SIZE, 7, 1, SpriteSheet.tiles, 16, 16);
-    public static final Sprite brick_exploded1 = new Sprite(DEFAULT_SIZE, 7, 2, SpriteSheet.tiles, 16, 16);
-    public static final Sprite brick_exploded2 = new Sprite(DEFAULT_SIZE, 7, 3, SpriteSheet.tiles, 16, 16);
-
-    /*
-    |--------------------------------------------------------------------------
-    | Powerups
-    |--------------------------------------------------------------------------
-     */
-    public static final Sprite powerup_bombs = new Sprite(DEFAULT_SIZE, 0, 10, SpriteSheet.tiles, 16, 16);
-    public static final Sprite powerup_flames = new Sprite(DEFAULT_SIZE, 1, 10, SpriteSheet.tiles, 16, 16);
-    public static final Sprite powerup_speed = new Sprite(DEFAULT_SIZE, 2, 10, SpriteSheet.tiles, 16, 16);
-    public static final Sprite powerup_wallpass = new Sprite(DEFAULT_SIZE, 3, 10, SpriteSheet.tiles, 16, 16);
-    public static final Sprite powerup_detonator = new Sprite(DEFAULT_SIZE, 4, 10, SpriteSheet.tiles, 16, 16);
-    public static final Sprite powerup_bombpass = new Sprite(DEFAULT_SIZE, 5, 10, SpriteSheet.tiles, 16, 16);
-    public static final Sprite powerup_flamepass = new Sprite(DEFAULT_SIZE, 6, 10, SpriteSheet.tiles, 16, 16);
+    private final int realWidth;
+    private final int realHeight;
+    private final SpriteSheet sheet;
 
     /**
      * Constructor.
      *
-     * @param size  size of sprite
-     * @param x     pos_X in sheet
-     * @param y     pos_y in sheet
-     * @param sheet sheet map image
-     * @param rw    real width
-     * @param rh    real height
+     * @param size       size of sprite
+     * @param crdX       tile coordinate in sheet
+     * @param crdY       tile coordinate in sheet
+     * @param sheet      sheet map image
+     * @param realWidth  real width
+     * @param realHeight real height
      */
-    public Sprite(int size, int x, int y, SpriteSheet sheet, int rw, int rh) {
+    public Sprite(int size, int crdX, int crdY, SpriteSheet sheet, int realWidth, int realHeight) {
         this.size = size;
-        this.pixels = new int[this.size * this.size];
-        this.x = x * this.size;
-        this.y = y * this.size;
+        this.x = crdX * this.size;
+        this.y = crdY * this.size;
         this.sheet = sheet;
-        this.realWidth = rw;
-        this.realHeight = rh;
+        this.realWidth = realWidth;
+        this.realHeight = realHeight;
+        this.pixels = new int[this.size * this.size];
 
         load();
     }
 
     /**
-     * Constructor.
-     *
-     * @param size  size of sprite.
-     * @param color color of sprite.
-     */
-    public Sprite(int size, int color) {
-        this.size = size;
-        pixels = new int[this.size * this.size];
-
-        setColor(color);
-    }
-
-    /**
-     * Fill all element in _pixels with given color.
-     *
-     * @param color color to fill
-     */
-    private void setColor(int color) {
-        Arrays.fill(pixels, color);
-    }
-
-    /**
-     * Copy image sheet area to sprite image area
+     * Copy image sheet area to sprite image area.
      */
     private void load() {
         for (int y = 0; y < size; y++) {
@@ -242,104 +48,61 @@ public final class Sprite {
     }
 
     /**
-     * Moving 1.
+     * Get pixel value at position (x,y).
      *
-     * @param normal  normal sprite
-     * @param x1      x1 sprite
-     * @param x2      x2 sprite
-     * @param animate animate
-     * @param time    time
-     * @return Sprite
+     * @param x posX
+     * @param y posY
+     * @return pixel value
      */
-    public static Sprite movingSprite(Sprite normal, Sprite x1, Sprite x2, int animate, int time) {
-        int calc = animate % time;
-        int diff = time / 3;
-
-        if (calc < diff) {
-            return normal;
-        }
-
-        if (calc < diff * 2) {
-            return x1;
-        }
-
-        return x2;
+    private int getPixel(int x, int y) {
+        return this.pixels[x + y * size];
     }
 
     /**
-     * Moving 2.
-     *
-     * @param x1      x1 sprite
-     * @param x2      x2 sprite
-     * @param animate animate
-     * @param time    time
-     * @return Sprite
-     */
-    public static Sprite movingSprite(Sprite x1, Sprite x2, int animate, int time) {
-        int diff = time / 2;
-        return (animate % time > diff) ? x1 : x2;
-    }
-
-    /**
-     * Copy Image.
+     * Get display pixel as Image Node.
      *
      * @return Image Node
      */
     public Image getFxImage() {
+        final int scaleFactor = SCALED_SIZE / DEFAULT_SIZE;
+
         // Create empty frame with given size
-        WritableImage wr = new WritableImage(this.size, this.size);
-        PixelWriter pw = wr.getPixelWriter();
+        WritableImage writableImage = new WritableImage(
+                this.size * scaleFactor,
+                this.size * scaleFactor);
+        PixelWriter pixelWriter = writableImage.getPixelWriter();
 
         // Copy image
         for (int x = 0; x < this.size; x++) {
             for (int y = 0; y < this.size; y++) {
-                if (pixels[x + y * size] == TRANSPARENT_COLOR) {
-                    pw.setArgb(x, y, 0);
-                } else {
-                    pw.setArgb(x, y, pixels[x + y * size]);
+                // read pixel and remove background color
+                final int pixel = this.getPixel(x, y);
+                final int finalPixel = pixel == TRANSPARENT_COLOR ? 0 : pixel;
+
+                // Fill pixel color
+                for (int dy = 0; dy < scaleFactor; dy++) {
+                    for (int dx = 0; dx < scaleFactor; dx++) {
+                        pixelWriter.setArgb(
+                                x * scaleFactor + dx,
+                                y * scaleFactor + dy,
+                                finalPixel);
+                    }
                 }
+
             }
         }
 
-        Image image = new ImageView(wr).getImage();
-
-        // Resize image
-        return resize(image, SCALED_SIZE / DEFAULT_SIZE);
+        return new ImageView(writableImage).getImage();
     }
 
     /**
-     * Resize.
+     * choose suitable sprite according to time passed
      *
-     * @param input       input image
-     * @param scaleFactor scale
-     * @return Image Node
+     * @param frameCount      time passed by frame unit
+     * @param framesForSprite total time frame unit for all sprite
+     * @param sprites         sprites
+     * @return suitable sprite
      */
-    private Image resize(Image input, int scaleFactor) {
-        final int W = (int) input.getWidth();
-        final int H = (int) input.getHeight();
-
-        WritableImage output = new WritableImage(
-                W * scaleFactor,
-                H * scaleFactor
-        );
-
-        PixelReader reader = input.getPixelReader();
-        PixelWriter writer = output.getPixelWriter();
-
-        for (int y = 0; y < H; y++) {
-            for (int x = 0; x < W; x++) {
-                final int argb = reader.getArgb(x, y);
-                for (int dy = 0; dy < scaleFactor; dy++) {
-                    for (int dx = 0; dx < scaleFactor; dx++) {
-                        writer.setArgb(x * scaleFactor + dx, y * scaleFactor + dy, argb);
-                    }
-                }
-            }
-        }
-
-        return output;
-    }
-
     @SafeVarargs
     public static <T> T selectSprite(int frameCount, int framesForSprite, T... sprites) {
         int numObject = sprites.length;
