+++ src/uet/oop/bomberman/sound/AudioFileName.java
@@ -1,0 +1,11 @@
+package uet.oop.bomberman.sound;
+
+/**
+ * Define audio file name.
+ */
+public interface AudioFileName {
+
+    String BG_SOUND = "bg_sound";
+    String BOM_EXPLODE = "bom_explode";
+    String BOM_SET = "bom_set";
+}
