+++ src/uet/oop/bomberman/untility/PathFile.java
@@ -3,6 +3,8 @@ package uet.oop.bomberman.untility;
 import uet.oop.bomberman.BombermanGame;
 
 import java.io.InputStream;
+import java.net.MalformedURLException;
+import java.net.URL;
 import java.util.Objects;
 
 public class PathFile {
@@ -11,6 +13,20 @@ public class PathFile {
     }
 
     public static String getPath(String path) {
-        return "res\\" + path;
+        try {
+            return Objects.requireNonNull(BombermanGame.class.getResource(path)).getPath();
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+        return "res" + path;
+    }
+
+    public static URL getURL(String path) throws MalformedURLException {
+        try {
+            return Objects.requireNonNull(BombermanGame.class.getResource(path));
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+        return new URL("file","", PathFile.getPath(path));
     }
 }
